added mediaBrowserServiceCompact(for getting connection to the media for selected caller and load UI contents)
