+++ app/src/main/java/com/raheemjnr/jr_music/utils/extraz.kt
@@ -22,3 +22,5 @@ fun askPermissions(context: Context, requestCode: Int, vararg permissions: Strin
         requestCode
     )
 }
+
+const val TAG = "MainActivityVM"
